add check if item is None

+++ src/arcgis/geoprocessing/_job.py
@@ -963,21 +963,24 @@ class OMJob(GPJob):
                 for key in item_keys:
                     if key in mission_json["items"].keys():
                         item_info = mission_json["items"][key]
-                        if "itemId" in item_info.keys():
+                        if isinstance(item_info, dict) and "itemId" in item_info.keys():
                             item_object = mission._gis.content.get(item_info["itemId"])
                             try:
                                 if item_object:
                                     deleted = item_object.delete()
                             except:
                                 pass
-                            mission_json["items"].update({key: {}})
+                        mission_json["items"].update({key: {}})
 
             if item_name == "dsm" or item_name == "dtm" or item_name == "ortho":
                 if "items" in mission_json.keys():
                     for key in mission_json["items"].keys():
                         if key == item_name:
                             item_info = mission_json["items"][key]
-                            if "itemId" in item_info.keys():
+                            if (
+                                isinstance(item_info, dict)
+                                and "itemId" in item_info.keys()
+                            ):
                                 item_object = mission._gis.content.get(
                                     item_info["itemId"]
                                 )
@@ -986,11 +989,9 @@ class OMJob(GPJob):
                                         deleted = item_object.delete()
                                 except:
                                     pass
-                                mission_json["items"].update({key: {}})
-                                if key in mission_json["jobs"].keys():
-                                    mission_json["jobs"].update(
-                                        {key: {"checked": False}}
-                                    )
+                            mission_json["items"].update({key: {}})
+                            if key in mission_json["jobs"].keys():
+                                mission_json["jobs"].update({key: {"checked": False}})
 
             if processing_states is not None:
                 mission_json["processingSettings"].update(
